Move BaseRequest methods into HandlesDeletion trait
