+++ src/Http/Requests/DestroyCategory.php
@@ -2,12 +2,16 @@
 
 namespace TeamTeaTime\Forum\Http\Requests;
 
+use Illuminate\Foundation\Http\FormRequest;
 use TeamTeaTime\Forum\Events\UserDeletedCategory;
 use TeamTeaTime\Forum\Interfaces\FulfillableRequest;
+use TeamTeaTime\Forum\Http\Requests\Traits\HandlesDeletion;
 use TeamTeaTime\Forum\Models\Category;
 
-class DestroyCategory extends BaseRequest implements FulfillableRequest
+class DestroyCategory extends FormRequest implements FulfillableRequest
 {
+    use HandlesDeletion;
+
     public function authorize(): bool
     {
         return $this->user()->can('delete', $this->route('category'));
