+++ src/Http/Requests/MoveThread.php
@@ -2,12 +2,13 @@
 
 namespace TeamTeaTime\Forum\Http\Requests;
 
+use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Support\Facades\DB;
 use TeamTeaTime\Forum\Events\UserMovedThread;
 use TeamTeaTime\Forum\Models\Category;
 use TeamTeaTime\Forum\Interfaces\FulfillableRequest;
 
-class MoveThread extends BaseRequest implements FulfillableRequest
+class MoveThread extends FormRequest implements FulfillableRequest
 {
     private Category $destinationCategory;
