+++ src/Http/Requests/PinThread.php
@@ -2,10 +2,11 @@
 
 namespace TeamTeaTime\Forum\Http\Requests;
 
+use Illuminate\Foundation\Http\FormRequest;
 use TeamTeaTime\Forum\Events\UserPinnedThread;
 use TeamTeaTime\Forum\Interfaces\FulfillableRequest;
 
-class PinThread extends BaseRequest implements FulfillableRequest
+class PinThread extends FormRequest implements FulfillableRequest
 {
     public function authorize(): bool
     {
