+++ src/Http/Requests/StoreCategory.php
@@ -2,11 +2,12 @@
 
 namespace TeamTeaTime\Forum\Http\Requests;
 
+use Illuminate\Foundation\Http\FormRequest;
 use TeamTeaTime\Forum\Events\UserCreatedCategory;
 use TeamTeaTime\Forum\Interfaces\FulfillableRequest;
 use TeamTeaTime\Forum\Models\Category;
 
-class StoreCategory extends BaseRequest implements FulfillableRequest
+class StoreCategory extends FormRequest implements FulfillableRequest
 {
     public function authorize(): bool
     {
